requirements.txt
discord.py
asyncio
dnspython
pymongo
pillow
easy-pil
discord-py-slash-command
discord-components
discord-slash-components-bridge
jishaku
